chore: move ChoiceJSPage into core/tests/pages.py
