+++ conftest.py
@@ -1,7 +1,5 @@
 import contextlib
 from datetime import datetime
-import functools
-from functools import cached_property
 import os
 import random
 from typing import Any, Iterable
@@ -14,7 +12,7 @@ from django.urls import resolve
 from django.urls.base import reverse
 from django.utils import timezone
 from django.utils.timezone import localdate
-from playwright.sync_api import Error as PlaywrightError, Page, expect
+from playwright.sync_api import Page, expect
 import pytest
 
 from core.constants import AC_STRUCTURE, DEPARTEMENTS, MUS_STRUCTURE
@@ -22,31 +20,6 @@ from core.factories import ContactStructureFactory, StructureFactory, UserFactor
 from core.models import Agent, Contact, Departement, Region, Structure
 
 User = get_user_model()
-
-
-def playwright_repeatable(maybe_func=None, timeout=None, retries=10):
-    def decorator(func):
-        @functools.wraps(func)
-        def wrapper(*args, **kwargs):
-            error = None
-            old_timeout = expect._timeout
-            new_timeout = timeout or (old_timeout or 5_000 / retries)
-            for i in range(retries):
-                try:
-                    expect.set_options(timeout=new_timeout)
-                    return func(*args, **kwargs)
-                except (PlaywrightError, AssertionError) as e:
-                    if error is not None:
-                        e.__cause__ = error
-                    error = e
-                    continue
-                finally:
-                    expect.set_options(timeout=old_timeout)
-            raise error
-
-        return wrapper
-
-    return decorator(maybe_func) if callable(maybe_func) else decorator
 
 
 class E2ETestNetworkError(Exception):
@@ -105,75 +78,6 @@ def mocked_authentification_user(db, request):
             yield user
 
 
-class _ChoiceJSPage:
-    @cached_property
-    def _choice_widget_locator(self):
-        locator = self.page.locator(self.sel_or_locator).locator(
-            "xpath=descendant-or-self::*[contains(concat(' ',normalize-space(@class),' '),' choices ')]"
-        )
-        if len(locator.all()) == 0:
-            return self.page.locator(self.sel_or_locator).locator(
-                "xpath=ancestor-or-self::*[contains(concat(' ',normalize-space(@class),' '),' choices ')]"
-            )
-        return locator
-
-    @property
-    def choice_widget(self):
-        return self._choice_widget_locator.first
-
-    @property
-    def dropdown(self):
-        return self.choice_widget.locator(".choices__list.choices__list--dropdown")
-
-    @property
-    def click_zone(self):
-        result = self.choice_widget.locator(".choices__list.choices__list--multiple + .choices__input")
-        return result.first if result.count() > 0 else self.choice_widget
-
-    def __init__(self, page: Page, sel_or_locator, fill_content, exact_name, *, check_selection=None):
-        self.page = page
-        self.sel_or_locator = sel_or_locator
-        self.fill_content = fill_content
-        self.exact_name = exact_name
-        self._check_selection = check_selection
-
-    def try_open(self):
-        if self.dropdown.is_visible():
-            return
-
-        self.click_zone.click()
-        expect(self.dropdown).to_be_visible()
-
-    def get_selected_option(self, exact_name):
-        return self.choice_widget.locator(
-            ".choices__list.choices__list--single,.choices__list.choices__list--multiple"
-        ).get_by_role("option", name=exact_name)
-
-    def select_option(self, exact_name):
-        self.choice_widget.locator(".choices__list.choices__list--dropdown").get_by_role(
-            "option", name=exact_name, exact=True
-        ).locator("visible=true").click()
-        expect(self.get_selected_option(exact_name)).to_have_count(1)
-
-    @playwright_repeatable
-    def try_select_option(self):
-        self.try_open()
-        if len(self.get_selected_option(self.exact_name).all()) > 0:
-            # Option already selected
-            return
-
-        self.choice_widget.locator("input.choices__input").fill(self.fill_content)
-        self.choice_widget.locator(".choices__list.choices__list--dropdown").get_by_role(
-            "option", name=self.exact_name, exact=True
-        ).locator("visible=true").click()
-        self.check_selection()
-
-    def check_selection(self):
-        if callable(self._check_selection):
-            return self._check_selection()
-        return expect(self.get_selected_option(self.exact_name)).to_have_count(1)
-
-
 @pytest.fixture
 def choice_js_fill():
     def _choice_js_fill(
@@ -185,7 +89,9 @@ def choice_js_fill():
         *,
         check_selection=None,
     ):
-        _ChoiceJSPage(
+        from core.tests.pages import ChoiceJSPage
+
+        ChoiceJSPage(
             page, sel_or_locator, fill_content, exact_name, check_selection=check_selection
         ).try_select_option()
